+++ docs/linker.md
@@ -2,43 +2,13 @@
 tags:
   - API
 ---
+
 # Documentation for `Linker` object
 
-
 ::: splink.linker.Linker
-    handler: python
-    selection:
-      members:
-        - __init__
-        - cluster_pairwise_predictions_at_threshold
-        - cluster_studio_dashboard
-        - compare_two_records
-        - comparison_viewer_dashboard
-        - compute_number_of_comparisons_generated_by_blocking_rule
-        - compute_tf_table
-        - debug_mode
-        - delete_tables_created_by_splink_from_db
-        - deterministic_link
-        - estimate_m_from_label_column
-        - estimate_parameters_using_expectation_maximisation
-        - estimate_u_using_random_sampling
-        - find_matches_to_new_records
-        - initialise_settings
-        - load_settings_from_json
-        - m_u_parameters_chart
-        - match_weight_histogram
-        - match_weights_chart
-        - missingness_chart
-        - parameter_estimate_comparisons_chart
-        - precision_recall_chart_from_labels
-        - predict
-        - profile_columns
-        - roc_chart_from_labels
-        - roc_table_from_labels
-        - save_settings_to_json
-        - train_m_from_pairwise_labels
-        - unlinkables_chart
-        - waterfall_chart
-    rendering:
-      show_root_heading: false
-      show_source: true
+handler: python
+selection:
+members: - **init** - cluster_pairwise_predictions_at_threshold - cluster_studio_dashboard - compare_two_records - comparison_viewer_dashboard - compute_number_of_comparisons_generated_by_blocking_rule - compute_tf_table - debug_mode - delete_tables_created_by_splink_from_db - deterministic_link - estimate_m_from_label_column - estimate_parameters_using_expectation_maximisation - estimate_u_using_random_sampling - find_matches_to_new_records - initialise_settings - load_settings_from_json - m_u_parameters_chart - match_weights_histogram - match_weights_chart - missingness_chart - parameter_estimate_comparisons_chart - precision_recall_chart_from_labels - predict - profile_columns - roc_chart_from_labels - roc_table_from_labels - save_settings_to_json - estimate_m_from_pairwise_labels - unlinkables_chart - waterfall_chart
+rendering:
+show_root_heading: false
+show_source: true
