+++ docs/linkerest.md
@@ -2,18 +2,13 @@
 tags:
   - API
 ---
+
 # Documentation for `Linker` object methods related to parameter estimation
 
-
 ::: splink.linker.Linker
-    handler: python
-    selection:
-      members:
-        - estimate_m_from_label_column
-        - estimate_parameters_using_expectation_maximisation
-        - estimate_u_using_random_sampling
-        - save_settings_to_json
-        - train_m_from_pairwise_labels
-    rendering:
-      show_root_heading: false
-      show_source: true
+handler: python
+selection:
+members: - estimate_m_from_label_column - estimate_parameters_using_expectation_maximisation - estimate_u_using_random_sampling - save_settings_to_json - estimate_m_from_pairwise_labels
+rendering:
+show_root_heading: false
+show_source: true
