+++ docs/linkerqa.md
@@ -2,24 +2,13 @@
 tags:
   - API
 ---
+
 # Documentation for `Linker` object methods related to QA
 
-
 ::: splink.linker.Linker
-    handler: python
-    selection:
-      members:
-        - cluster_studio_dashboard
-        - comparison_viewer_dashboard
-        - m_u_parameters_chart
-        - match_weight_histogram
-        - match_weights_chart
-        - parameter_estimate_comparisons_chart
-        - precision_recall_chart_from_labels
-        - roc_chart_from_labels
-        - roc_table_from_labels
-        - unlinkables_chart
-        - waterfall_chart
-    rendering:
-      show_root_heading: false
-      show_source: true
+handler: python
+selection:
+members: - cluster_studio_dashboard - comparison_viewer_dashboard - m_u_parameters_chart - match_weights_histogram - match_weights_chart - parameter_estimate_comparisons_chart - precision_recall_chart_from_labels - roc_chart_from_labels - roc_table_from_labels - unlinkables_chart - waterfall_chart
+rendering:
+show_root_heading: false
+show_source: true
