+++ splink/charts.py
@@ -21,6 +21,7 @@ try:
 
         return VegaliteNoValidate(spec)
 
+
 except ImportError:
     altair_installed = False
 
@@ -240,8 +241,8 @@ def precision_recall_chart(records, width=400, height=400, as_dict=False):
     return vegalite_or_json(chart, as_dict=as_dict)
 
 
-def match_weight_histogram(records, width=500, height=250, as_dict=False):
-    chart_path = "match_weight_histogram.json"
+def match_weights_histogram(records, width=500, height=250, as_dict=False):
+    chart_path = "match_weights_histogram.json"
     chart = load_chart_definition(chart_path)
 
     chart["data"]["values"] = records
