+++ tests/test_m_train.py
@@ -56,7 +56,7 @@ def test_m_train():
     linker_pairwise = DuckDBLinker(df, settings)
 
     linker_pairwise._con.register("labels", df_labels)
-    linker_pairwise.train_m_from_pairwise_labels("labels")
+    linker_pairwise.estimate_m_from_pairwise_labels("labels")
     cc_name = linker_pairwise._settings_obj.comparisons[0]
 
     cl_exact = cc_name._get_comparison_level_by_comparison_vector_value(2)
